feat(mps-sync-plugin-lib): extract model, module and node transformers to classes
