+++ mps-sync-plugin-lib/src/main/kotlin/org/modelix/mps/sync/SyncService.kt
@@ -5,6 +5,7 @@ import org.modelix.kotlin.utils.UnstableModelixFeature
 import org.modelix.model.api.INode
 import org.modelix.model.client2.ModelClientV2
 import org.modelix.model.lazy.BranchReference
+import org.modelix.model.mpsadapters.MPSLanguageRepository
 
 @UnstableModelixFeature(reason = "The new modelix MPS plugin is under construction", intendedFinalization = "2024.1")
 interface SyncService {
@@ -12,9 +13,9 @@ interface SyncService {
     suspend fun bindModel(
         client: ModelClientV2,
         branchReference: BranchReference,
-        modelName: String,
         model: INode,
         targetProject: MPSProject,
+        languageRepository: MPSLanguageRepository,
         afterActivate: (() -> Unit)?,
     ): IBinding
 }
