+++ mps-sync-plugin-lib/src/main/kotlin/org/modelix/mps/sync/mps/factories/SNodeFactory.kt
@@ -153,6 +153,8 @@ class SNodeFactory(
             }
         }
     }
+
+    fun clearResolvableReferences() = resolvableReferences.clear()
 }
 
 @UnstableModelixFeature(reason = "The new modelix MPS plugin is under construction", intendedFinalization = "2024.1")
